third_party/WebKit/LayoutTests/http/tests/security/suborigins/suborigin-unsafe-postmessage-send.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>Validate that unsafe-postmessage-send allows Suborigin to send messages as physical origin via postMessage.</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<script>
async_test(t => {
    window.addEventListener('message', make_test(t, 'iframe'));
  },
  'Validate serialization of event.origin and event.suborigin in ' +
  'postMessage from an iframe using \'unsafe-postmessage-send\'');
async_test(t => {
    window.addEventListener('message', make_test(t, 'window'));
  },
  'Validate serialization of event.origin and event.suborigin in ' +
  'postMessage from a window using \'unsafe-postmessage-send\'');

function make_test(t, name) {
  return t.step_func(function(event) {
      if (event.data.type === name) {
        assert_equals(event.origin, 'http://127.0.0.1:8000');
        assert_equals(event.suborigin, name);
        assert_equals(event.data.suborigin,  name);
        t.done();
      }
    });
}

window.open(
  'resources/post-document-to-parent.php?suborigin=window&type=window&' +
  'suboriginpolicy=\'unsafe-postmessage-send\'');
</script>
<iframe src="resources/post-document-to-parent.php?suborigin=iframe&type=iframe&suboriginpolicy='unsafe-postmessage-send'"></iframe>
</html>
